% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ini.R
\name{generate_ini}
\alias{generate_ini}
\title{Gera um arquivo de autenticação .ini}
\usage{
generate_ini(url, usr, pwd)
}
\arguments{
\item{url}{String com a url base de acesso ao servidor}

\item{usr}{String com a identificação do usuário}

\item{pwd}{String com a senha de acesso do usuário}
}
\value{
Não existe nenhum retorno para a função.
}
\description{
Função para criar um arquivo de autenticação com os dados fornecidos
}
\details{
O arquivo será gerado com uma única seção, login, que é por padrão a que todas funções neste pacote utilizam.
O local de criação será o mesmo de sua R Session, e irá conter o nome auth.ini .
}
\examples{
generate_ini("https://4intelligence.com.br/", "example@4i.com.br", "example_pwd")


}
\author{
João Gustavo Oliveira
}
