% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_r_environ.R
\name{generate_r_environ}
\alias{generate_r_environ}
\title{Gera um arquivo para armazenar as variáveis de autenticação}
\usage{
generate_r_environ(url, usr, pwd)
}
\arguments{
\item{url}{String com a url base de acesso ao servidor}

\item{usr}{String com a identificação do usuário}

\item{pwd}{String com a senha de acesso do usuário}
}
\value{
Não existe nenhum retorno para a função.
}
\description{
Função para criar ou adicionar linhas (caso já exista) no arquivo **.Renviron** com os dados fornecidos
}
\details{
O arquivo será gerado com uma única seção, login, que é por padrão a que todas funções neste pacote utilizam.
O local de criação será o mesmo de sua R Session, e irá conter o nome **.Renviron** .

Para acessar o valor das credenciais, utilize a função `Sys.getenv()` e passe como parâmetro a variável que deseja.

Os valores disponíveis são:
`Sys.getenv(URL_4MACRO)`
`Sys.getenv(USR_4MACRO)`
`Sys.getenv(PWD_4MACRO)`
}
\examples{
generate_r_environ("https://4intelligence.com.br/", "example@4i.com.br", "example_pwd")

}
\author{
Samuel Souza
}
