% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimated_labels.R
\name{get_estimated_labels}
\alias{get_estimated_labels}
\title{Consulta Labels de Projeções}
\usage{
get_estimated_labels(serie, ...)
}
\arguments{
\item{serie}{String com o código de 16 dígitos da série}

\item{...}{Parâmetros adicionais: \code{filepath} String com caminho para o arquivo \code{.ini} de autenticação}
}
\value{
O retorno consiste em uma lista contendo todos os labels de projeção para a série requisitada
}
\description{
Função para consultar todos os labels de projeção de uma determinada série
}
\details{
Primeiramente é necessário salvar as suas credenciais em um arquivo \code{.ini} ou \code{.Renviron}
Para gerar um arquivo de autenticação utilize a função \code{generate_ini} ou \code{generate_r_environ}
É altamente recomendável utilizar o \code{.Renviron} como seu arquivo de autenticação.

O parâmetro obrigatório, \code{serie}, deve ser o código de 16 dígitos de uma série com projeção existente
}
\examples{
\dontrun{
# Utilizando um arquivo .ini

get_estimated_labels(serie = 'BRGDP0002000ROQL', filepath = "/home/user/auth.ini")

# Utilizando variáveis de ambiente

get_estimated_labels(serie = 'BRGDP0002000ROQL')

}

}
\author{
João Gustavo Oliveira, Gabriel Belle
}
