% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_series.R
\name{insert_series}
\alias{insert_series}
\title{Insere Valores em uma Série Especificada}
\usage{
insert_series(
  filepath,
  serie,
  overwrite,
  access_group,
  contents,
  estimate,
  label_estimate = NULL
)
}
\arguments{
\item{filepath}{String com caminho para o arquivo \code{.ini} de autenticação, para gerar um arquivo utilize a função \code{generate_ini}}

\item{serie}{String com o código de 16 digitos da série}

\item{overwrite}{Logical para definir se as observações poderão ser sobrescritas ou não}

\item{access_group}{String com o nome do grupo de acesso para inserir a série}

\item{contents}{DataFrame contendo as observações com obrigatóriamente duas colunas: \code{date} e \code{val}, as datas devem estar no formato ISO \code{yyyy-mm-dd}}

\item{estimate}{Logical \code{TRUE} se são dados projetados, para dados realizados \code{FALSE}}

\item{label_estimate}{String com nome da projeção, obrigatório somente para dados projetados}
}
\value{
O retorno consiste em uma lista contendo dois campos, code e message, code contém o código da resposta do servidor e
message um texto explicitando o código recebido. Os possíveis códigos de retorno são:

\item{0}{Dados inseridos com sucesso;}
\item{1}{Erro no formato da requisição;}
\item{2}{Parâmetro usr não está no formato correto;}
\item{3}{Parâmetro pwd não está no formato correto;}
\item{4}{Código da série não está no formato correto;}
\item{5}{Grupo de acesso não está no formato correto;}
\item{6}{Parâmetro estimate não está no formato correto;}
\item{7}{Parâmetro label_estimate não está no formato correto;}
\item{8}{Data frame com os dados não está no formato correto;}
\item{9}{Erro de autenticação com os dados do usuário;}
\item{10}{Grupo de acesso não está disponível para o usuário;}
\item{11}{Série não existe para o grupo de acesso;}
\item{12}{Valores do data frame não estão no formato correto;}
\item{13}{Dados duplicados, neste caso um terceiro valor será retornado, repeated_date, indicando os dados duplicados.}
}
\description{
Função para inserir dados estimados ou realizados em uma determinada série
}
\details{
O arquivo de autenticação \code{.ini} deve conter uma única seção com o nome \code{login}, incluindo o
seguintes campos:

\itemize{
\item{url: }{Url base de acesso ao servidor;}
\item{usr: }{Id do usuário;}
\item{pwd: }{Senha do usuário.}
}


Para criar um arquivo de autenticação utilize a função \code{generate_ini}
}
\examples{
\dontrun{
df <- data.frame("date" = c("2020-05-13"), "val" = c("21.5"))
insert_series("/home/joao/auth.ini", "BRLDG0002000SOML", TRUE, "Geral", df, FALSE)
insert_series("/home/joao/auth.ini", "BRLDG0002000SOML", FALSE, "Geral", df, TRUE, 
    "Estimado em 2020-05-13")
}

}
\author{
João Gustavo Oliveira
}
